move storage classes to new directory
