+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/src/Internals/PersistentStorage/AzureMonitorPersistentStorage.cs
@@ -4,7 +4,7 @@
 using System.Diagnostics;
 using System.Threading;
 
-namespace Azure.Monitor.OpenTelemetry.Exporter.Internals
+namespace Azure.Monitor.OpenTelemetry.Exporter.Internals.PersistentStorage
 {
     internal class AzureMonitorPersistentStorage
     {
